Add tentative hill climbers
This is still a not-very clean or stable implementation.
`hill_climbing.py` itself should crack a substitution cipher when run.

`autoperm_hill_climbing.py` should spectacularly fail to do anything.
Also, autoperm_hill_climbing.py is v e r y   s l o w. This is in part
due to Python just being slow, in part due to the nature of the cipher,
and in part because of the abstract, class-heavy way everything is
implemented here.